requirements.txt
flask==2.3.3
transformers==4.41.1
torch==2.3.0
faiss-cpu==1.8.0
sentence-transformers==2.6.1
scikit-learn==1.4.2
pymongo==4.7.2
numpy==1.26.4
nltk==3.8.1
python-dotenv==1.0.1
